requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
gunicorn        
psycopg2-binary 